- cleaned up a whole pile of compiler warnings
git-svn-id: https://megaglest.svn.sourceforge.net/svnroot/megaglest/trunk@1126 03bcafef-9c4d-45c2-aa05-caeccb650a8a

+++ source/shared_lib/sources/graphics/gl/shader_gl.cpp
@@ -68,7 +68,7 @@ bool ShaderProgramGl::link(string &messages){
 	assertGl();
 
 	//bind attributes
-	for(int i=0; i<attributes.size(); ++i){
+	for(unsigned int i=0; i<attributes.size(); ++i){
 		int a= attributes[i].second;
 		string s= attributes[i].first;
 		glBindAttribLocationARB(handle, attributes[i].second, attributes[i].first.c_str());
@@ -186,7 +186,7 @@ bool ShaderGl::compile(string &messages){
 	messages= "Compiling shader: " + source.getPathInfo() + "\n";
 
 	//load source
-	GLint length= source.getCode().size();
+	GLint length= (GLint)source.getCode().size();
 	const GLcharARB *csource= source.getCode().c_str();
 	glShaderSourceARB(handle, 1, &csource, &length);
